pyproject.toml
[tool.poetry]
name = "echo-crafter"
version = "0.1.0"
description = "Unleash the power of thought-driven developement"
authors = ["Jean-Francois Arbour <jf.arbour@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.12,<3.13"
openai = "^1.6.1"
pvrecorder = "^1.2.2"
pvporcupine = "^3.0.2"
pvrhino = "^3.0.2"
pvcobra = "^2.0.2"
pvleopard = "^2.0.2"
psutil = "^5.9.8"
numpy = "^1.26.3"
tabulate = "^0.9.0"
rich = "^13.7.0"
prompt-toolkit = "^3.0.43"
pyaml = "^23.12.0"
pvcheetah = "^2.0.1"

[tool.poetry.group.dev.dependencies]
pyright = "^1.1.352"
python-language-server = "^0.36.2"
black = "^24.2.0"
aider-chat = "^0.25.0"


[tool.poetry.group.docs.dependencies]
sphinx = "^7.2.6"

[tool.poetry.scripts]
restart_daemons = "scripts.restart_daemons:main"
make_scripts = "echo_crafter.prompts.make-script:main"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
